[FEAT] 23/09/21 취미게시판 최종 취합 전 커밋
기능 최종수정 및 CSS 수정

+++ src/main/java/com/alone/special/hobby/controller/ReplyController.java
@@ -32,42 +32,43 @@ public class ReplyController {
 	@RequestMapping(value="/insert.do", method=RequestMethod.POST)
 	public ModelAndView insertReply(ModelAndView mv
 			, @ModelAttribute Reply reply
-			, @RequestParam("category") String refCategoryName
+			, @ModelAttribute Board board
 			, HttpSession session) {
 		
 		String url = "";
 		try {
-//			String replyWriter = (String)session.getAttribute("memberId");		// 로그인 되어있는 아이디로 댓글 등록되게
-//			reply.sethReplyWriter(replyWriter);
-			int result = rService.insertReply(reply);
-			
-			Board board = new Board();
-			board.setRefCategoryName(refCategoryName);
 			String encodedCategory = URLEncoder.encode(board.getRefCategoryName(), "UTF-8");
 			url = "/hobby/board/detail.do?category=" + encodedCategory + "&hBoardNo=" + reply.getRefBoardNo();
 			
-			if(result > 0) {
-				int refBoardNo = reply.getRefBoardNo();
-				int replyTotalCount = rService.getReplyCount(reply.getRefBoardNo());
+			String replyWriter = (String)session.getAttribute("userId");
+			if(replyWriter != null && !replyWriter.equals("")) {
+				reply.sethReplyWriter(replyWriter);
+				int result = rService.insertReply(reply);
 				
-				Map<String, Integer> replyCountInfo = new HashMap<>();
-				
-		        replyCountInfo.put("refBoardNo", refBoardNo);
-		        replyCountInfo.put("replyTotalCount", replyTotalCount);
-				
-				result = bService.updateReplyNum(replyCountInfo);
-				mv.setViewName("redirect:"+url);
+				if(result > 0) {
+					int refBoardNo = reply.getRefBoardNo();
+					int replyTotalCount = rService.getReplyCount(reply.getRefBoardNo());
+					
+					Map<String, Integer> replyCountInfo = new HashMap<>();
+			        replyCountInfo.put("refBoardNo", refBoardNo);
+			        replyCountInfo.put("replyTotalCount", replyTotalCount);
+					
+					result = bService.updateReplyNum(replyCountInfo);
+					mv.setViewName("redirect:"+url);
+				} else {
+					mv.addObject("msg", "댓글 등록이 완료되지 않았습니다.");
+					mv.addObject("url", url);
+					mv.setViewName("common/errorPage");
+				}
 			} else {
-				mv.addObject("msg", "댓글 등록이 완료되지 않았습니다.");
-				mv.addObject("error", "댓글 등록 실패");
+				mv.addObject("msg", "로그인이 완료되지 않았습니다.");
 				mv.addObject("url", url);
 				mv.setViewName("common/errorPage");
 			}
 			
 		} catch (Exception e) {
 			mv.addObject("msg", "관리자에게 문의하세요.");
-			mv.addObject("error", e.getMessage());
-			mv.addObject("url", url);		// reply에서 가져온 외래키 boardNo 사용
+			mv.addObject("url", url);
 			mv.setViewName("common/errorPage");
 		}
 		return mv;
@@ -76,33 +77,37 @@ public class ReplyController {
 	@RequestMapping(value="/update.do", method = RequestMethod.POST)
 	public ModelAndView updateReply(ModelAndView mv
 			, @ModelAttribute Reply reply
-			, @RequestParam("category") String refCategoryName
+			, @ModelAttribute Board board
 			, HttpSession session) {
 
 		String url = "";
 		try {
-//			String replyWriter = (String)session.getAttribute("memberId");
-//			if(!replyWriter.equals("")) {
-//				reply.sethReplyWriter(replyWriter);
+			String encodedCategory = URLEncoder.encode(board.getRefCategoryName(), "UTF-8");
+			url = "/hobby/board/detail.do?category=" + encodedCategory + "&hBoardNo=" + reply.getRefBoardNo();
 			
-				Board board = new Board();
-				board.setRefCategoryName(refCategoryName);
-				String encodedCategory = URLEncoder.encode(board.getRefCategoryName(), "UTF-8");
-				url = "/hobby/board/detail.do?category=" + encodedCategory + "&hBoardNo=" + reply.getRefBoardNo();
-				
-				int result = rService.updateReply(reply);
-				mv.setViewName("redirect:"+url);
-//			} else {
-//				mv.addObject("msg", "로그인이 완료되지 않았습니다.");
-//				mv.addObject("error", "로그인 정보 확인 요망");
-//				mv.addObject("url", "/index.jsp");
-//				mv.setViewName("common/errorPage");
-//			}
+			String memberId = (String)session.getAttribute("userId");
+			String replyWriter = reply.gethReplyWriter();
+			if(memberId != null && !memberId.equals("")) {
+				if(replyWriter.equals(memberId)) {
+					int result = rService.updateReply(reply);
+					if(result > 0) {
+						mv.setViewName("redirect:"+url);
+					}
+					mv.setViewName("redirect:"+url);
+				} else {
+					mv.addObject("msg", "자신의 댓글만 수정 가능합니다.");
+					mv.addObject("url", url);
+					mv.setViewName("common/errorPage");
+				}
+			} else {
+				mv.addObject("msg", "로그인이 완료되지 않았습니다.");
+				mv.addObject("url", url);
+				mv.setViewName("common/errorPage");
+			}
 			
 		} catch (Exception e) {
 			mv.addObject("msg", "관리자에게 문의하세요.");
-			mv.addObject("error", e.getMessage());
-			mv.addObject("url", url);		// reply에서 가져온 외래키 boardNo 사용
+			mv.addObject("url", url);
 			mv.setViewName("common/errorPage");
 		}
 		return mv;
@@ -113,35 +118,40 @@ public class ReplyController {
 			, @ModelAttribute Reply reply
 			, @RequestParam("category") String refCategoryName
 			, HttpSession session) {
+		
 		String url = "";
 		try {
-//			String memberId = (String)session.getAttribute("memberId");
-//			String replyWriter = reply.gethReplyWriter();
 			
 			Board board = new Board();
 			board.setRefCategoryName(refCategoryName);
 			String encodedCategory = URLEncoder.encode(board.getRefCategoryName(), "UTF-8");
 			url = "/hobby/board/detail.do?category=" + encodedCategory + "&hBoardNo=" + reply.getRefBoardNo();
-//			if(replyWriter != null && replyWriter.equals(memberId)) {	// 세션 아이디, 작성자 비교
-				int result = rService.deleteReply(reply);
-				if(result > 0) {
-					mv.setViewName("redirect:"+url);
+			
+			String memberId = (String)session.getAttribute("userId");
+			if(memberId != null && !memberId.equals("")) {
+				String replyWriter = reply.gethReplyWriter();
+				if(replyWriter.equals(memberId)) {
+					int result = rService.deleteReply(reply);
+					if(result > 0) {
+						mv.setViewName("redirect:"+url);
+					} else {
+						mv.addObject("msg", "댓글 삭제가 완료되지 않았습니다.");
+						mv.addObject("url", url);
+						mv.setViewName("common/errorPage");
+					}
 				} else {
-					mv.addObject("msg", "댓글 삭제가 완료되지 않았습니다.");
-					mv.addObject("error", "댓글 삭제 실패");
+					mv.addObject("msg", "자신의 댓글만 삭제 가능합니다.");
 					mv.addObject("url", url);
 					mv.setViewName("common/errorPage");
 				}
-//			} else {
-//				mv.addObject("msg", "자신의 댓글만 삭제 가능합니다.");
-//				mv.addObject("error", "댓글 삭제 불가");
-//				mv.addObject("url", url);
-//				mv.setViewName("common/errorPage");
-//			}
-				
+			} else {
+				mv.addObject("msg", "로그인이 완료되지 않았습니다.");
+				mv.addObject("url", url);
+				mv.setViewName("common/errorPage");
+			}
+			
 		} catch (Exception e) {
 			mv.addObject("msg", "관리자에게 문의하세요.");
-			mv.addObject("error", e.getMessage());
 			mv.addObject("url", url);
 			mv.setViewName("common/errorPage");
 		}
